refactor: 适配新的 default 组织
